Example.jl Documentation

julia> using GaussianBeamPropagation
julia> using Unitful # optional, but enables the use of the @u_str macros
julia> λ = 1030u"nm"1030 nm
julia> w0_in = 2.75u"mm"2.75 mm
julia> beam_in = Beam(λ;
           wx=w0_in,
           wy=w0_in,
           rx=Inf * u"m",
           ry=Inf * u"m",
       )Beam{Float64, SecondOrderMoments{Float64}}(1.03e-6, SecondOrderMoments{Float64}(1.890625e-6, 0.0, 1.890625e-6, 3.553442767806815e-9, 0.0, 3.553442767806815e-9, 0.0, 0.0, 0.0, 0.0))
julia> M = FreeSpace(0.5) *
               ThinLens(OpticalPower(1 / 0.5)) *
               FreeSpace(0.5)MatrixElement([0.0 0.0 0.5 0.0; 0.0 0.0 0.0 0.5; -2.0 0.0 0.0 0.0; 0.0 -2.0 0.0 0.0])
julia> beam_out = M * beam_inBeam{Float64, StaticArraysCore.SMatrix{4, 4, Float64, 16}}(1.03e-6, [8.883606919517038e-10 0.0 0.0 0.0; 0.0 8.883606919517038e-10 0.0 0.0; 0.0 0.0 7.5625e-6 0.0; 0.0 0.0 0.0 7.5625e-6])
julia> radius(beam_out)(x = 5.961076050350989e-5, y = 5.961076050350989e-5, φ = 0.0)
julia> phase_curvature(beam_out)(x = -Inf, y = -Inf, φ = 0.0)

GaussianBeamPropagation.AbstractElement — Type

AbstractElement

TODO

GaussianBeamPropagation.FreeSpace — Type

FreeSpace

TODO

GaussianBeamPropagation.MatrixElement — Type

MatrixElement

TODO

GaussianBeamPropagation.Mirror — Type

Mirror(surface::Curvature; incidence::IncidenceAngle)
Mirror(power::OpticalPower; incidence::IncidenceAngle)

Matrix representing a mirror with surface curvature surface or an optical power of power. The incident direction of the beam is given by the optional parameter incidence.

GaussianBeamPropagation.OpticalInterface — Type

OpticalInterface(surface::Curvature, n::Real)
OpticalInterface(surface::Curvature, n::Real, incidence::IncidenceAngle)

Matrix representing an optical interface

This is a matrix builder that constructs the matrix for a refractive interface between two materials.

The arguments are the following:

surface::Curvature: curvature of the interface given by Curvature, where a positive curvature denotes that the center of curvature lies after the interface.

n::Real: ratio of the refractive index of the material after the interface to the refractive index of the material before the interface.

incidence::IncidenceAngle (optional): if given, defines the direction of the incident ray or beam using IncidenceAngle. If absent, it is assumed that the ray is parallel to the surface normal in the center (disregarding any surface curvature).

Coordinate system

The coordinate system used by the matrix is right handed, where the ray or beam travels in the +z direction. Both the orientation angle of the surface and the orientation of the incidence plane are rotations in the x-y-plane relative to the x̂ direction.

GaussianBeamPropagation.ThinLens — Type

ThinLens(focussing::Union{FocalLength,FocalPower} [, n::Real] [, ia::IncidenceAngle])
ThinLens(front::RadiusOfCurvature, back::RadiusOfCurvature [, n::Real] [, ia::IncidenceAngle])

The zero-thickness approximation of the thick lens. The oblique astigmatism depends on the refractive indices of the lens material and the surrounding medium, which are given as a single ratio of internal / external, and the incidence angles are given by IncidenceAngle. The focussing of the lens can either be given as RadiusOfCurvature, FocalLength, or FocalPower.

GaussianBeamPropagation.ThickLens — Function

ThickLens(front_surface::Curvature, thickness::Real, back_surface::Curvature, n::Real)
ThickLens(front_surface::Curvature, thickness::Real, back_surface::Curvature, n::Real, incidence::IncidenceAngle)

Matrix representing a thick lens

This is a matrix builder that constructs the matrix for a thick lens.

The arguments are the following:

front_surface::Curvature: curvature of the front lens surface given by Curvature, where a positive curvature denotes that the center of curvature lies after the lens, i.e., the front surface is convex.

thickness::Real: thickness of the lens where the ray or beam passes through.

back_surface::Curvature: curvature of the back given by Curvature, where a positive curvature denotes that the center of curvature lies after the lens, i.e., the back surface is concave.

n::Real: ratio of the refractive index of the lens material to the refractive index of the surrounding material.

incidence::IncidenceAngle (optional): if given, defines the direction of the incident ray or beam using IncidenceAngle. If absent, it is assumed that the ray is parallel to the surface normal in the center (disregarding any surface curvature).

Coordinate system

The coordinate system used by the matrix is right handed, where the ray or beam travels in the +z direction. Both the orientation angle of the surface and the orientation of the incidence plane are rotations in the x-y-plane relative to the x̂ direction.

GaussianBeamPropagation.Curvature — Type

Curvature(κ::Real)
Curvature(κx::Real, κy::Real; φ::Real)
Curvature(κ_xx::Real, κ_xy::Real, κ_yy::Real)

Spherical curvature

Defines the curvature of a spherical surface from a single curvature κ, which is defined as the reciprocal radius of curvature R, such that κ = R⁻¹. It has a single argument:

κ::Real: defines the spherical curvature of the surface.

Toroidal curvature

To define a toroidal curvature (including a cylindrical surface, two radii of curvature in two principal axes are required:

κx::Real: defines the surface curvature in the x' direction.

κy::Real: defines the surface curvature in the y' direction.

φ::Real (keyword, optional): defines the rotation of the x' and y' directions relative to the global x and y directions.

Projected curvature

Instead of providing the curvature in the principal directions with a rotation angle, you can provide the components of the vectors already rotated:

κ_xx::Real: projected curvature in the x direction.

κ_xy::Real: (κx - κy) * cos(φ) * sin(φ)

κ_yy::Real: projected curvature in the y direction

Coordinate system

The coordinate system used by the matrix is right handed, where the ray or beam travels in the +z direction. Both the orientation angle of the surface and the orientation of the incidence plane are rotations in the x-y-plane relative to the x̂ direction.

The curvature κ of a surface is the inverse of its radius of curvature R, such that R ≡ κ⁻¹. It is given in units of inverse metre.

GaussianBeamPropagation.IncidenceAngle — Type

IncidenceAngle(ι, θ)

Description of an incidence direction different from the surface normal of the mirror, lens or interface. It is given by an incidence angle ι, the angle between the incident direction and the surface normal, and the rotation angle θ giving the orientation of the incidence plane relative to the x-direction in a right-handed coordinate system.