Elements

AbstractElement

All elements are subtypes of AbstractElement. It is used only for functions that apply to each Element.

GaussianBeamPropagation.AbstractElement — Type

AbstractElement

Supertype of all elements that can be used in the Wigner formalism. This includes optical interfaces, free space propagation, and other elements that can be represented by a transfer matrix.

Real Elements

GaussianBeamPropagation.FreeSpace — Type

FreeSpace(d::Real)

Matrix representing a propagation of the beam through free space.

The arguments are the following:

d::Real: propagation distance

GaussianBeamPropagation.MatrixElement — Type

MatrixElement(matrix::SMatrix{4,4})

Element that is defined by an arbitrary transfer matrix.

GaussianBeamPropagation.Mirror — Type

Mirror(surface::Curvature)
Mirror(surface::Curvature, incidence::IncidenceAngle)
Mirror(power::OpticalPower)
Mirror(power::OpticalPower, incidence::IncidenceAngle)

Matrix representing a mirror with surface curvature surface or an optical power of power. The incident direction of the beam is given by the optional parameter incidence.

GaussianBeamPropagation.OpticalInterface — Type

OpticalInterface(surface::Curvature, n::Real)
OpticalInterface(surface::Curvature, n::Real, incidence::IncidenceAngle)

Matrix representing an optical interface

This is a matrix builder that constructs the matrix for a refractive interface between two materials.

The arguments are the following:

surface::Curvature: curvature of the interface given by Curvature, where a positive curvature denotes that the center of curvature lies after the interface.

n::Real: ratio of the refractive index of the material after the interface to the refractive index of the material before the interface.

incidence::IncidenceAngle (optional): if given, defines the direction of the incident ray or beam using IncidenceAngle. If absent, it is assumed that the ray is parallel to the surface normal in the center (disregarding any surface curvature).

Coordinate system

The coordinate system used by the matrix is right handed, where the ray or beam travels in the +z direction. Both the orientation angle of the surface and the orientation of the incidence plane are rotations in the x-y-plane relative to the x̂ direction.

GaussianBeamPropagation.ThinLens — Type

ThinLens(power::OpticalPower)
ThinLens(power::OpticalPower, n::Real, incidence::IncidenceAngle)
ThinLens(front::Curvature, back::Curvature, n::Real)
ThinLens(front::Curvature, back::Curvature, n::Real, incidence::IncidenceAngle)

The zero-thickness approximation of the thick lens. The oblique astigmatism depends on the refractive indices of the lens material and the surrounding medium, which are given as a single ratio of internal / external, and the incidence angles are given by IncidenceAngle. The focussing of the lens can either be given as RadiusOfCurvature, FocalLength, or FocalPower.

GaussianBeamPropagation.ThickLens — Function

ThickLens(front_surface::Curvature, thickness::Real, back_surface::Curvature, n::Real)
ThickLens(front_surface::Curvature, thickness::Real, back_surface::Curvature, n::Real, incidence::IncidenceAngle)

Matrix representing a thick lens

This is a matrix builder that constructs the matrix for a thick lens.

The arguments are the following:

front_surface::Curvature: curvature of the front lens surface given by Curvature, where a positive curvature denotes that the center of curvature lies after the lens, i.e., the front surface is convex.

thickness::Real: thickness of the lens where the ray or beam passes through.

back_surface::Curvature: curvature of the back given by Curvature, where a positive curvature denotes that the center of curvature lies after the lens, i.e., the back surface is concave.

n::Real: ratio of the refractive index of the lens material to the refractive index of the surrounding material.

incidence::IncidenceAngle (optional): if given, defines the direction of the incident ray or beam using IncidenceAngle. If absent, it is assumed that the ray is parallel to the surface normal in the center (disregarding any surface curvature).

Coordinate system

The coordinate system used by the matrix is right handed, where the ray or beam travels in the +z direction. Both the orientation angle of the surface and the orientation of the incidence plane are rotations in the x-y-plane relative to the x̂ direction.