API Reference

GaussianBeamPropagation.ThickLens — Function

ThickLens(front_surface::Curvature, thickness::Real, back_surface::Curvature, n::Real)
ThickLens(front_surface::Curvature, thickness::Real, back_surface::Curvature, n::Real, incidence::IncidenceAngle)

Matrix representing a thick lens

This is a matrix builder that constructs the matrix for a thick lens.

The arguments are the following:

front_surface::Curvature: curvature of the front lens surface given by Curvature, where a positive curvature denotes that the center of curvature lies after the lens, i.e., the front surface is convex.

thickness::Real: thickness of the lens where the ray or beam passes through.

back_surface::Curvature: curvature of the back given by Curvature, where a positive curvature denotes that the center of curvature lies after the lens, i.e., the back surface is concave.

n::Real: ratio of the refractive index of the lens material to the refractive index of the surrounding material.

incidence::IncidenceAngle (optional): if given, defines the direction of the incident ray or beam using IncidenceAngle. If absent, it is assumed that the ray is parallel to the surface normal in the center (disregarding any surface curvature).

Coordinate system

The coordinate system used by the matrix is right handed, where the ray or beam travels in the +z direction. Both the orientation angle of the surface and the orientation of the incidence plane are rotations in the x-y-plane relative to the x̂ direction.

GaussianBeamPropagation.divergence — Method

divergence(::Beam)
divergence(::SecondOrderMoments)

Calculates the divergence angle from a Beam or SecondOrderMoments. The return value is a NamedTuple with entries (:x, :y, :φ), where both x and y are given in radians, and the angle φ is given in radians as well.

It uses a right-handed coordinate system.

GaussianBeamPropagation.phase_curvature — Method

phase_curvature(::Beam)
phase_curvature(::SecondOrderMoments)

Calculates the phase-front curvature from a Beam or SecondOrderMoments. The return value is a NamedTuple with entries (:x, :y, :φ), where both x and y are given in metres/radians, and the angle φ is given in radians.

It uses a right-handed coordinate system.

GaussianBeamPropagation.radius — Method

radius(::Beam)
radius(::SecondOrderMoments)
radius(xx, xy, yy)

Calculates the 1/e² radius from a Beam, SecondOrderMoments or measured second order moments (xx, xy, and yy). The return value is a NamedTuple with entries (:x, :y, :φ), where both x and y are given in metres, and the angle φ is given in radians. It follows the definition of ISO 11146 for x, y, and the angle φ.

It uses a right-handed coordinate system.

GaussianBeamPropagation.twist — Method

twist(::Beam)
twist(::SecondOrderMoments)

Calculates the twist from a Beam or SecondOrderMoments. The return value is a single Float64 in units of metres²/radians².

GaussianBeamPropagation.AbstractElement — Type

AbstractElement

Supertype of all elements that can be used in the Wigner formalism. This includes optical interfaces, free space propagation, and other elements that can be represented by a transfer matrix.

GaussianBeamPropagation.Curvature — Type

Curvature(κ::Real)
Curvature(κx::Real, κy::Real; φ::Real)
Curvature(κ_xx::Real, κ_xy::Real, κ_yy::Real)

Spherical curvature

Defines the curvature of a spherical surface from a single curvature κ, which is defined as the reciprocal radius of curvature R, such that κ = R⁻¹. It has a single argument:

κ::Real: defines the spherical curvature of the surface.

Toroidal curvature

To define a toroidal curvature (including a cylindrical surface, two radii of curvature in two principal axes are required:

κx::Real: defines the surface curvature in the x' direction.

κy::Real: defines the surface curvature in the y' direction.

φ::Real (keyword, optional): defines the rotation of the x' and y' directions relative to the global x and y directions.

Projected curvature

Instead of providing the curvature in the principal directions with a rotation angle, you can provide the components of the vectors already rotated:

κ_xx::Real: projected curvature in the x direction.

κ_xy::Real: (κx - κy) * cos(φ) * sin(φ)

κ_yy::Real: projected curvature in the y direction

Coordinate system

The coordinate system used by the matrix is right handed, where the ray or beam travels in the +z direction. Both the orientation angle of the surface and the orientation of the incidence plane are rotations in the x-y-plane relative to the x̂ direction.

The curvature κ of a surface is the inverse of its radius of curvature R, such that R ≡ κ⁻¹. It is given in units of inverse metre.

GaussianBeamPropagation.FreeSpace — Type

FreeSpace(d::Real)

Matrix representing a propagation of the beam through free space.

The arguments are the following:

d::Real: propagation distance

GaussianBeamPropagation.IncidenceAngle — Type

IncidenceAngle(ι, θ)

Description of an incidence direction different from the surface normal of the mirror, lens or interface. It is given by an incidence angle ι, the angle between the incident direction and the surface normal, and the rotation angle θ giving the orientation of the incidence plane relative to the x-direction in a right-handed coordinate system.

GaussianBeamPropagation.MatrixElement — Type

MatrixElement(matrix::SMatrix{4,4})

Element that is defined by an arbitrary transfer matrix.

GaussianBeamPropagation.Mirror — Type

Mirror(surface::Curvature)
Mirror(surface::Curvature, incidence::IncidenceAngle)
Mirror(power::OpticalPower)
Mirror(power::OpticalPower, incidence::IncidenceAngle)

Matrix representing a mirror with surface curvature surface or an optical power of power. The incident direction of the beam is given by the optional parameter incidence.

GaussianBeamPropagation.OpticalInterface — Type

OpticalInterface(surface::Curvature, n::Real)
OpticalInterface(surface::Curvature, n::Real, incidence::IncidenceAngle)

Matrix representing an optical interface

This is a matrix builder that constructs the matrix for a refractive interface between two materials.

The arguments are the following:

surface::Curvature: curvature of the interface given by Curvature, where a positive curvature denotes that the center of curvature lies after the interface.

n::Real: ratio of the refractive index of the material after the interface to the refractive index of the material before the interface.

incidence::IncidenceAngle (optional): if given, defines the direction of the incident ray or beam using IncidenceAngle. If absent, it is assumed that the ray is parallel to the surface normal in the center (disregarding any surface curvature).

Coordinate system

The coordinate system used by the matrix is right handed, where the ray or beam travels in the +z direction. Both the orientation angle of the surface and the orientation of the incidence plane are rotations in the x-y-plane relative to the x̂ direction.

GaussianBeamPropagation.OpticalPower — Type

OpticalPower(P)
OpticalPower(Px, Py; φ)
OpticalPower(P_xx, P_xy, P_yy)

The optical power P of a lens or mirror describes the change in convergence or divergence of light, equal to the reciprocal focal length f such that P ≡ f⁻¹. It is given in units of inverse metre.

GaussianBeamPropagation.SecondOrderMoments — Type

SecondOrderMoments(; 
    rxrx, rxry, ryry,
    θxθx, θxθy, θyθy,
    rxθx, rxθy, ryθx, ryθy,
)

GaussianBeamPropagation.ThinLens — Type

ThinLens(power::OpticalPower)
ThinLens(power::OpticalPower, n::Real, incidence::IncidenceAngle)
ThinLens(front::Curvature, back::Curvature, n::Real)
ThinLens(front::Curvature, back::Curvature, n::Real, incidence::IncidenceAngle)

The zero-thickness approximation of the thick lens. The oblique astigmatism depends on the refractive indices of the lens material and the surrounding medium, which are given as a single ratio of internal / external, and the incidence angles are given by IncidenceAngle. The focussing of the lens can either be given as RadiusOfCurvature, FocalLength, or FocalPower.